Fix pixel alignment when using transform tool on pixel art
There were two problems:

1) KisTransformWorker accepted the offset measured in integers,
   which created a small drift of the whole image

2) KisFilterWeightsApplicator calculated the destination rect
   too aggressively. The formula might be okay for the filters
   that have whole-pixel support, but for half-pixel-support
   filters it caused issues.

BUG:413668

+++ plugins/tools/tool_transform2/kis_transform_utils.cpp
@@ -255,8 +255,8 @@ KisTransformWorker KisTransformUtils::createTransformWorker(const ToolTransformA
                                        config.originalCenter().x(),
                                        config.originalCenter().y(),
                                        config.aZ(),
-                                       (int)(translation.x()),
-                                       (int)(translation.y()),
+                                       translation.x(),
+                                       translation.y(),
                                        updater,
                                        config.filter());
